migrate variants test to new test framework
