--- crates/test-rust-wasm/src/bin/variants.rs
@@ -1,3 +1,0 @@
-include!("../../../../tests/runtime/variants/wasm.rs");
-
-fn main() {}
