+++ tests/runtime/main.rs
@@ -29,7 +29,6 @@ mod resource_with_lists;
 mod resources;
 mod results;
 mod rust_xcrate;
-mod variants;
 mod versions;
 
 struct MyCtx {}
